+++ docs/src/examples/components/Segment/Variations/index.tsx
@@ -9,6 +9,11 @@ const Variations = () => (
       description="A segment can have its colors inverted for contrast."
       examplePath="components/Segment/Variations/SegmentExampleInverted"
     />
+    <ComponentExample
+      title="Color"
+      description="A segment can have different colors."
+      examplePath="components/Segment/Variations/SegmentExampleColor"
+    />
   </ExampleSection>
 )